move: FakeRepository移動
